move grocery ground tasks to teacher_tasks.py
